% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_clean.R
\docType{data}
\name{music_clean}
\alias{music_clean}
\title{Cleaned version of Music Genre Classification 'train' data set (renamed 'music')}
\format{
## `music_clean`
A data frame with 11,813 rows and 15 columns:
\describe{
  \item{Popularity}{Track popularity based on all-time streams (0 to 100)}
  \item{danceability}{Track suitability for dancing (0.0 to 1.0)}
  \item{energy}{The perceptual measure of intensity and activity (0.0 to 1.0)}
  \item{key}{The key the track is in (-1 to -11)}
  \item{loudness}{Overall loudness of the track in decibels (-60 to 0 dB)}
  \item{mode}{Modality of the track (Major'1'/Minor'0')}
  \item{speechiness}{Presence of spoken words in the track (0.0 to 1.0)}
  \item{acousticness}{Confidence measure from 0 to 1 of whether the track is acoustic}
  \item{instrumentalness}{The likelihood the track contains vocal content (0.0 to 1.0}
  \item{liveness}{Presence of audience in the recordings (0.0 to 1.0)}
  \item{valence}{Musical positiveness (0.0 to 1.0)}
  \item{tempo}{Tempo of the track in beats per minute (BPM)}
  \item{duration_in min/ms}{Duration of the track in minutes (min) or milliseconds (ms)}
  \item{time_signature}{Estimated time signature (3 to 7)}
  \item{Class}{Genre: Rock, Indie, Alt, Pop, Metal, HipHop, Alt_Music, Blues, Acoustic/Folk, Instrumental, Country, or Bollywood}
  ...
}
}
\source{
<https://www.kaggle.com/datasets/purumalgi/music-genre-classification/data?select=train.csv>
}
\usage{
music_clean
}
\description{
A subset of data from the Music Genre Classification data set on Kaggle
Created during a MachineHack Hackathon, presumably from music data provided via APIs on Spotify
Audio features described here: https://www.kaggle.com/datasets/amitanshjoshi/spotify-1million-tracks?resource=download
}
\details{
Cleaning code to remove qualitative variables, NAs, and relabel grouping variable 'Class' (hide from analysis)
music_clean <- music |> dplyr::select(-"Artist Name", -"Track Name") |> tidyr::drop_na()
music_clean$Class <- as.factor(music_clean$Class)
levels(music_clean$Class) <- c("Rock", "Indie", "Alt", "Pop", "Metal", "Hip Hop", "Alt_Music", "Blues", "Acoustic/Folk", "Instrumental", "Country")
}
\keyword{datasets}
